requirements.txt
# CHRONO — Dependencies
# Core
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.7.0
python-multipart==0.0.9
numpy>=1.24.0
pandas>=2.0.0

# Image Processing
opencv-python-headless==4.10.0.84
pillow>=10.3.0

# AI / Gemma 4
requests>=2.32.0
torch>=2.3.0
transformers>=4.48.0
accelerate>=0.30.0
huggingface-hub>=0.23.0

# Agent
langchain>=0.2.0
langgraph>=0.1.0

# Fine-Tuning (Unsloth Track)
# unsloth  # Install separately: pip install unsloth
# trl>=0.9.0
# peft>=0.11.0

# Testing
pytest>=8.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
